# src/main/java/com/example/spring_security_jwt/entity/User.java
package com.example.spring_security_jwt.entity;

import java.sql.Timestamp;

import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import jakarta.persistence.Id;
import jakarta.persistence.Table;
import lombok.Data;

@Table(name = "users")
@Entity
@Data
public class User {

	@Id
	@GeneratedValue(strategy = GenerationType.IDENTITY)
	private Long id;
	private String username;
	private String email;
	private String password;
	private String fullName;
	private boolean isActive;
	private Timestamp createdAt;
	private Timestamp updatedAt;

}